app-server/requirements.txt
paho-mqtt
numpy
requests
# pymongo
aiohttp